Clean up check style issues
