+++ core/cas-server-core-logout/src/main/java/org/apereo/cas/logout/config/CasCoreLogoutConfiguration.java
@@ -4,17 +4,18 @@ import org.apache.commons.lang3.StringUtils;
 import org.apereo.cas.authentication.AuthenticationServiceSelectionPlan;
 import org.apereo.cas.configuration.CasConfigurationProperties;
 import org.apereo.cas.logout.DefaultLogoutExecutionPlan;
+import org.apereo.cas.logout.DefaultLogoutManager;
 import org.apereo.cas.logout.DefaultSingleLogoutServiceLogoutUrlBuilder;
 import org.apereo.cas.logout.DefaultSingleLogoutServiceMessageHandler;
 import org.apereo.cas.logout.LogoutExecutionPlan;
 import org.apereo.cas.logout.LogoutExecutionPlanConfigurer;
 import org.apereo.cas.logout.LogoutManager;
-import org.apereo.cas.logout.DefaultLogoutManager;
 import org.apereo.cas.logout.LogoutMessageCreator;
 import org.apereo.cas.logout.SamlCompliantLogoutMessageCreator;
 import org.apereo.cas.logout.SingleLogoutServiceLogoutUrlBuilder;
 import org.apereo.cas.logout.SingleLogoutServiceMessageHandler;
 import org.apereo.cas.services.ServicesManager;
+import org.apereo.cas.ticket.registry.TicketRegistry;
 import org.apereo.cas.util.http.HttpClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -38,6 +39,10 @@ import java.util.List;
 @EnableConfigurationProperties(CasConfigurationProperties.class)
 public class CasCoreLogoutConfiguration implements LogoutExecutionPlanConfigurer {
     private static final Logger LOGGER = LoggerFactory.getLogger(CasCoreLogoutConfiguration.class);
+
+    @Autowired
+    @Qualifier("ticketRegistry")
+    private TicketRegistry ticketRegistry;
 
     @Autowired
     private CasConfigurationProperties casProperties;
@@ -76,7 +81,7 @@ public class CasCoreLogoutConfiguration implements LogoutExecutionPlanConfigurer
     @Autowired
     @Bean
     public LogoutManager logoutManager(@Qualifier("logoutExecutionPlan") final LogoutExecutionPlan logoutExecutionPlan) {
-        return new DefaultLogoutManager(logoutBuilder(), defaultSingleLogoutServiceMessageHandler(), 
+        return new DefaultLogoutManager(logoutBuilder(), defaultSingleLogoutServiceMessageHandler(),
                 casProperties.getSlo().isDisabled(), logoutExecutionPlan);
     }
 
@@ -97,5 +102,15 @@ public class CasCoreLogoutConfiguration implements LogoutExecutionPlanConfigurer
             c.configureLogoutExecutionPlan(plan);
         });
         return plan;
+    }
+
+    @Override
+    public void configureLogoutExecutionPlan(final LogoutExecutionPlan plan) {
+        plan.registerLogoutHandler(ticketGrantingTicket -> ticketGrantingTicket.getDescendantTickets()
+                .stream()
+                .forEach(t -> {
+                    LOGGER.debug("Deleting ticket [{}] from the registry as a descendant of [{}]", t, ticketGrantingTicket.getId());
+                    ticketRegistry.deleteTicket(t);
+                }));
     }
 }
