+++ support/cas-server-support-oauth/src/main/java/org/apereo/cas/config/CasOAuthAuthenticationServiceSelectionStrategyConfiguration.java
@@ -2,18 +2,13 @@ package org.apereo.cas.config;
 
 import org.apereo.cas.authentication.AuthenticationServiceSelectionPlan;
 import org.apereo.cas.authentication.AuthenticationServiceSelectionStrategy;
+import org.apereo.cas.authentication.AuthenticationServiceSelectionStrategyConfigurer;
 import org.apereo.cas.authentication.principal.ServiceFactory;
 import org.apereo.cas.authentication.principal.WebApplicationService;
-import org.apereo.cas.authentication.AuthenticationServiceSelectionStrategyConfigurer;
 import org.apereo.cas.configuration.CasConfigurationProperties;
-import org.apereo.cas.logout.LogoutExecutionPlan;
-import org.apereo.cas.logout.LogoutExecutionPlanConfigurer;
-import org.apereo.cas.logout.LogoutHandler;
 import org.apereo.cas.services.ServicesManager;
 import org.apereo.cas.support.oauth.util.OAuth20Utils;
 import org.apereo.cas.support.oauth.validator.OAuth20AuthenticationServiceSelectionStrategy;
-import org.apereo.cas.support.oauth.logout.OAuth20LogoutHandler;
-import org.apereo.cas.ticket.registry.TicketRegistry;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -30,12 +25,7 @@ import org.springframework.context.annotation.Configuration;
  */
 @Configuration("casOAuthAuthenticationServiceSelectionStrategyConfiguration")
 @EnableConfigurationProperties(CasConfigurationProperties.class)
-public class CasOAuthAuthenticationServiceSelectionStrategyConfiguration
-        implements AuthenticationServiceSelectionStrategyConfigurer, LogoutExecutionPlanConfigurer {
-
-    @Autowired
-    @Qualifier("ticketRegistry")
-    private TicketRegistry ticketRegistry;
+public class CasOAuthAuthenticationServiceSelectionStrategyConfiguration implements AuthenticationServiceSelectionStrategyConfigurer {
     
     @Autowired
     private CasConfigurationProperties casProperties;
@@ -55,21 +45,9 @@ public class CasOAuthAuthenticationServiceSelectionStrategyConfiguration
         return new OAuth20AuthenticationServiceSelectionStrategy(servicesManager,
                 webApplicationServiceFactory, OAuth20Utils.casOAuthCallbackUrl(casProperties.getServer().getPrefix()));
     }
-
-    @Bean
-    @ConditionalOnMissingBean(name = "oauth20LogoutHandler")
-    @RefreshScope
-    public LogoutHandler oauth20LogoutHandler() {
-        return new OAuth20LogoutHandler(servicesManager, ticketRegistry);
-    }
-
+    
     @Override
     public void configureAuthenticationServiceSelectionStrategy(final AuthenticationServiceSelectionPlan plan) {
         plan.registerStrategy(oauth20AuthenticationRequestServiceSelectionStrategy());
-    }
-
-    @Override
-    public void configureLogoutExecutionPlan(final LogoutExecutionPlan plan) {
-        plan.registerLogoutHandler(oauth20LogoutHandler());
     }
 }
