+++ support/cas-server-support-oauth/src/main/java/org/apereo/cas/support/oauth/web/response/accesstoken/AccessTokenRefreshTokenGrantRequestExtractor.java
@@ -6,7 +6,6 @@ import org.apereo.cas.support.oauth.OAuthConstants;
 import org.apereo.cas.support.oauth.util.OAuth20Utils;
 import org.apereo.cas.ticket.OAuthToken;
 import org.apereo.cas.ticket.registry.TicketRegistry;
-import org.pac4j.core.context.J2EContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
